Fixed the asyncs and added the finalExam page.
Managed to solve the async task problem by implementing a listener to be
called when the async task is complete. This allows me to return my data
to the calling activity without stopping the ui thread. Also managed to do
up the final exam page. Had another async task problem there; I needed to
call an async task for each module, so I made the activity pingpong between
two checking functions until I received data for all the modules. Not sure
if that is a good way to solve. Big update this time because I didnt commit
after I fixed the async tasks.

Signed-off-by: nicholaslum444 <nicholaslum444@hotmail.com>