Announcements, Finals, Gradebook are quite functional now.
Added refresh buttons to the announcement, finals and gradebook pages. Also
made alert dialogs popup when an item is clicked, e.g. announcement item to
show the announcement content and further details. Not a very big update,
but quite important as these functionalities are kinda essential. Finally,
added dummy classes for the homework, tests, and calculator.

Signed-off-by: nicholaslum444 <nicholaslum444@hotmail.com>