Fixed a few stuff, kinda ready for filming.
Added the missing About and Help pages. Removed the Friends option in the
side drawer. Changed the announcement fetching to only get the announcements
that are unread (used another api call). Because of this, also added a check
when entering the announcement page to refresh the modules info when entering
from the notification. Few other minor changes, lost track already.

Signed-off-by: nicholaslum444 <nicholaslum444@hotmail.com>