Fixed final exams and cap calculator display bugs.
Final exams:
The timings given did not include the time, only the date. Thus the time is
always 0:00 so I removed the time field in the display to prevent confusion.

Cap calc:
The current cap was given as -1 if there is no saved cap. Originally showed
the -1 value on the edittext. This is confusing, so now the cap only shows
if the value is not -1.

Signed-off-by: nicholaslum444 <nicholaslum444@hotmail.com>